Fixed compilation errors
- Added proper sidebar to views
- Added temp button for changing pages
